+++ HISTORY.rst
@@ -42,6 +42,8 @@ Unreleased Changes
   Pygeoprocessing >= 1.6.1. The HRA Proprocessor is removed and its previous
   functionality was simplified and merged into the HRA model itself.
   The model will no longer generate HTML plots and tables.
+* Add a software update notification button, dialog, and a link to the download
+  page on the User Interface when a new InVEST version is available.
 
 3.6.0 (2019-01-30)
 ------------------